Refactored code

+++ Cmds/delete.js
@@ -3,7 +3,15 @@ const Discord = require("discord.js");
 module.exports.run = async (bot, msg, args) => {
 
   let nDelete = args[0];
-  msg.channel.bulkDelete(nDelete); 
+  try {
+    if (nDelete > 100) {
+      msg.channel.send("The max you can delete at any given time is 1000.")
+    } else {
+      await msg.channel.bulkDelete(nDelete);
+    } 
+  } catch(err) {
+    console.error(err);
+  }
 
 }
